+++ node/middlewares/generateMiddlewares/generateSitemap.ts
@@ -29,10 +29,12 @@ const DEFAULT_REWRITER_ROUTES_PAYLOAD = {
 
 export async function generateSitemap(ctx: EventContext) {
   const { clients: { events }, body: { generationId }, state: { settings } }  = ctx
+  const disableDraftRoutes = settings.disableDraftRoutes
   if (settings.enableNavigationRoutes) {
     events.sendEvent('', GENERATE_REWRITER_ROUTES_EVENT, {
       ...DEFAULT_REWRITER_ROUTES_PAYLOAD,
       generationId,
+      disableDraftRoutes,
     } as RewriterRoutesGenerationEvent)
   }